Mobile UI: clear picking slots, prompt on overpick & fix toast notications (#18965)
* Mobile UI: clear picking slots, prompt on overpick & fix toast notifications
refs:
https://github.com/metasfresh/metasfresh/issues/18962
https://github.com/metasfresh/metasfresh/issues/18963
https://github.com/metasfresh/metasfresh/issues/18964
