+++ backend/de.metas.adempiere.adempiere/base/src/main/java-gen/org/compiere/model/X_MobileUI_UserProfile_Picking.java
@@ -12,7 +12,7 @@ import java.util.Properties;
 public class X_MobileUI_UserProfile_Picking extends org.compiere.model.PO implements I_MobileUI_UserProfile_Picking, org.compiere.model.I_Persistent 
 {
 
-	private static final long serialVersionUID = -678002359L;
+	private static final long serialVersionUID = 895391193L;
 
     /** Standard Constructor */
     public X_MobileUI_UserProfile_Picking (final Properties ctx, final int MobileUI_UserProfile_Picking_ID, @Nullable final String trxName)
@@ -141,6 +141,18 @@ public class X_MobileUI_UserProfile_Picking extends org.compiere.model.PO implem
 	public boolean isPickingWithNewLU()
 	{
 		return get_ValueAsBoolean(COLUMNNAME_IsPickingWithNewLU);
+	}
+
+	@Override
+	public void setIsShowConfirmationPromptWhenOverPick (final boolean IsShowConfirmationPromptWhenOverPick)
+	{
+		set_Value (COLUMNNAME_IsShowConfirmationPromptWhenOverPick, IsShowConfirmationPromptWhenOverPick);
+	}
+
+	@Override
+	public boolean isShowConfirmationPromptWhenOverPick() 
+	{
+		return get_ValueAsBoolean(COLUMNNAME_IsShowConfirmationPromptWhenOverPick);
 	}
 
 	@Override
