+++ backend/de.metas.handlingunits.base/src/main/java/de/metas/handlingunits/picking/config/MobileUIPickingUserProfile.java
@@ -75,6 +75,7 @@ public class MobileUIPickingUserProfile
 	boolean isCatchWeightTUPickingEnabled;
 	boolean considerSalesOrderCapacity;
 	boolean isAllowSkippingRejectedReason;
+	boolean isShowConfirmationPromptWhenOverPick;
 	@NonNull CreateShipmentPolicy createShipmentPolicy;
 	@Getter(AccessLevel.NONE) @NonNull PickingFiltersList filters;
 	@Getter(AccessLevel.PACKAGE) @NonNull ImmutableList<PickingJobField> fields;
@@ -93,6 +94,7 @@ public class MobileUIPickingUserProfile
 			final boolean isCatchWeightTUPickingEnabled,
 			final boolean considerSalesOrderCapacity,
 			final boolean isAllowSkippingRejectedReason,
+			final boolean isShowConfirmationPromptWhenOverPick,
 			final @Nullable CreateShipmentPolicy createShipmentPolicy,
 			final @Nullable PickingFiltersList filters,
 			final @NonNull ImmutableList<PickingJobField> fields)
@@ -122,6 +124,7 @@ public class MobileUIPickingUserProfile
 		this.considerSalesOrderCapacity = considerSalesOrderCapacity;
 		this.isAllowSkippingRejectedReason = isAllowSkippingRejectedReason;
 		this.isAllowNewTU = isAllowNewTU;
+		this.isShowConfirmationPromptWhenOverPick = isShowConfirmationPromptWhenOverPick;
 	}
 
 	public ImmutableList<PickingJobFacetGroup> getFilterGroupsInOrder() {return filters.getGroupsInOrder();}
