+++ backend/de.metas.handlingunits.base/src/main/java/de/metas/handlingunits/picking/job/repository/PickingJobLoaderAndSaver.java
@@ -1,11 +1,9 @@
 package de.metas.handlingunits.picking.job.repository;
 
-import com.google.common.collect.ArrayListMultimap;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.ImmutableSetMultimap;
-import com.google.common.collect.Maps;
 import com.google.common.collect.SetMultimap;
 import com.google.common.collect.Sets;
 import de.metas.bpartner.BPartnerId;
@@ -57,45 +55,32 @@ import de.metas.quantity.Quantity;
 import de.metas.quantity.Quantitys;
 import de.metas.uom.UomId;
 import de.metas.user.UserId;
-import de.metas.util.Check;
-import de.metas.util.GuavaCollectors;
 import de.metas.util.OptionalBoolean;
-import de.metas.util.Services;
 import lombok.NonNull;
-import org.adempiere.ad.dao.IQueryBL;
 import org.adempiere.exceptions.AdempiereException;
-import org.adempiere.model.InterfaceWrapperHelper;
 import org.adempiere.warehouse.LocatorId;
 
 import javax.annotation.Nullable;
-import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
-import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Stream;
 
-class PickingJobLoaderAndSaver
+class PickingJobLoaderAndSaver extends PickingJobSaver
 {
-	private final IQueryBL queryBL = Services.get(IQueryBL.class);
-	private final PickingJobLoaderSupportingServices _loadingSupportingServices;
+	@NonNull
+	private final PickingJobLoaderSupportingServices loadingSupportingServices;
 
-	private final HashMap<PickingJobId, I_M_Picking_Job> pickingJobs = new HashMap<>();
-	private final ArrayListMultimap<PickingJobId, I_M_Picking_Job_HUAlternative> pickingJobHUAlternatives = ArrayListMultimap.create();
-	private final ArrayListMultimap<PickingJobId, I_M_Picking_Job_Line> pickingJobLines = ArrayListMultimap.create();
-	private final ArrayListMultimap<PickingJobLineId, I_M_Picking_Job_Step> pickingJobSteps = ArrayListMultimap.create();
-	private final ArrayListMultimap<PickingJobStepId, I_M_Picking_Job_Step_HUAlternative> pickingJobStepAlternatives = ArrayListMultimap.create();
-	private final ArrayListMultimap<PickingJobStepId, I_M_Picking_Job_Step_PickedHU> pickedHUs = ArrayListMultimap.create();
 	private final HashMap<PickingJobId, Boolean> hasLocks = new HashMap<>();
 
-	private PickingJobLoaderAndSaver(@Nullable final PickingJobLoaderSupportingServices loadingSupportingServices)
+	private PickingJobLoaderAndSaver(@NonNull final PickingJobLoaderSupportingServices loadingSupportingServices)
 	{
-		this._loadingSupportingServices = loadingSupportingServices;
+		this.loadingSupportingServices = loadingSupportingServices;
 	}
 
 	public static PickingJobLoaderAndSaver forLoading(@NonNull final PickingJobLoaderSupportingServices loadingSupportingServices)
@@ -103,16 +88,16 @@ class PickingJobLoaderAndSaver
 		return new PickingJobLoaderAndSaver(loadingSupportingServices);
 	}
 
-	public static PickingJobLoaderAndSaver forSaving()
+	public static PickingJobSaver forSaving()
 	{
-		return new PickingJobLoaderAndSaver(null);
+		return new PickingJobSaver();
 	}
 
 	public PickingJob loadById(@NonNull final PickingJobId pickingJobId)
 	{
 		if (pickingJobs.get(pickingJobId) == null)
 		{
-			loadRecordsFromDB(ImmutableSet.of(pickingJobId));
+			loadRecords(ImmutableSet.of(pickingJobId));
 		}
 
 		return loadJob(pickingJobs.get(pickingJobId));
@@ -136,178 +121,6 @@ class PickingJobLoaderAndSaver
 				.map(pickingJobs::get)
 				.map(this::loadJob)
 				.collect(ImmutableList.toImmutableList());
-	}
-
-	public void save(@NonNull final PickingJob pickingJob)
-	{
-		final PickingJobId pickingJobId = pickingJob.getId();
-		if (pickingJobs.get(pickingJobId) == null)
-		{
-			loadRecordsFromDB(ImmutableSet.of(pickingJobId));
-		}
-
-		final I_M_Picking_Job record = pickingJobs.get(pickingJobId);
-		updateRecord(record, pickingJob);
-		InterfaceWrapperHelper.save(record);
-
-		final OrgId orgId = OrgId.ofRepoId(record.getAD_Org_ID());
-		saveLines(pickingJob.getLines(), pickingJobId, orgId, pickingJob.getDocStatus());
-	}
-
-	private void saveLines(
-			final ImmutableList<PickingJobLine> lines,
-			final PickingJobId pickingJobId,
-			final OrgId orgId,
-			final PickingJobDocStatus docStatus)
-	{
-		final HashMap<PickingJobLineId, I_M_Picking_Job_Line> existingRecords = pickingJobLines.get(pickingJobId)
-				.stream()
-				.collect(GuavaCollectors.toHashMapByKey(line -> PickingJobLineId.ofRepoId(line.getM_Picking_Job_Line_ID())));
-
-		for (final PickingJobLine line : lines)
-		{
-			final I_M_Picking_Job_Line existingRecord = existingRecords.get(line.getId());
-			Check.assumeNotNull(existingRecord, "line record shall exist for {}", line);
-
-			// NOTE: atm we have nothing to sync on line level
-			updateRecord(existingRecord, line, docStatus);
-			InterfaceWrapperHelper.save(existingRecord);
-
-			saveSteps(line.getSteps(), pickingJobId, line.getId(), orgId, docStatus);
-		}
-	}
-
-	private void saveSteps(
-			final ImmutableList<PickingJobStep> steps,
-			final PickingJobId pickingJobId,
-			final PickingJobLineId pickingJobLineId,
-			final OrgId orgId,
-			final PickingJobDocStatus docStatus)
-	{
-		final HashMap<PickingJobStepId, I_M_Picking_Job_Step> existingRecords = pickingJobSteps.get(pickingJobLineId)
-				.stream()
-				.collect(GuavaCollectors.toHashMapByKey(step -> PickingJobStepId.ofRepoId(step.getM_Picking_Job_Step_ID())));
-
-		for (final PickingJobStep step : steps)
-		{
-			final PickingJobStepPickFrom mainPickFrom = step.getPickFrom(PickingJobStepPickFromKey.MAIN);
-
-			I_M_Picking_Job_Step stepRecord = existingRecords.remove(step.getId());
-			if (stepRecord == null)
-			{
-				stepRecord = InterfaceWrapperHelper.newInstance(I_M_Picking_Job_Step.class);
-				stepRecord.setM_Picking_Job_Step_ID(step.getId().getRepoId());
-				stepRecord.setM_Picking_Job_ID(pickingJobId.getRepoId());
-				stepRecord.setM_Picking_Job_Line_ID(pickingJobLineId.getRepoId());
-				stepRecord.setIsDynamic(step.isGeneratedOnFly());
-				stepRecord.setAD_Org_ID(orgId.getRepoId());
-				stepRecord.setM_ShipmentSchedule_ID(step.getShipmentScheduleId().getRepoId());
-				stepRecord.setC_Order_ID(step.getSalesOrderAndLineId().getOrderRepoId());
-				stepRecord.setC_OrderLine_ID(step.getSalesOrderAndLineId().getOrderLineRepoId());
-
-				//
-				// What?
-				stepRecord.setM_Product_ID(step.getProductId().getRepoId());
-				stepRecord.setC_UOM_ID(step.getQtyToPick().getUomId().getRepoId());
-				stepRecord.setQtyToPick(step.getQtyToPick().toBigDecimal());
-			}
-
-			updateRecord(stepRecord, step, docStatus);
-			InterfaceWrapperHelper.save(stepRecord);
-
-			saveStepPickFromAlternatives(step, pickingJobId);
-
-			saveStepPickedTo(
-					mainPickFrom.getPickedTo(),
-					pickingJobId,
-					step.getId(),
-					null);
-		}
-
-		deleteStepsCascade(existingRecords.values());
-	}
-
-	private void saveStepPickFromAlternatives(
-			@NonNull final PickingJobStep step,
-			@NonNull final PickingJobId pickingJobId)
-	{
-		final ImmutableMap<PickingJobStepPickFromKey, I_M_Picking_Job_Step_HUAlternative> records = Maps.uniqueIndex(
-				pickingJobStepAlternatives.get(step.getId()),
-				record -> PickingJobStepPickFromKey.alternative(extractAlternativeId(record)));
-
-		for (final PickingJobStepPickFromKey pickFromKey : step.getPickFromKeys())
-		{
-			if (pickFromKey.isAlternative())
-			{
-				final PickingJobStepPickFrom pickFrom = step.getPickFrom(pickFromKey);
-				final I_M_Picking_Job_Step_HUAlternative record = Objects.requireNonNull(records.get(pickFromKey));
-
-				updateRecord(record, pickFrom);
-				InterfaceWrapperHelper.save(record);
-
-				saveStepPickedTo(
-						pickFrom.getPickedTo(),
-						pickingJobId,
-						step.getId(),
-						pickFromKey.getAlternativeId());
-			}
-		}
-	}
-
-	private void saveStepPickedTo(
-			@Nullable final PickingJobStepPickedTo pickedTo,
-			@NonNull final PickingJobId pickingJobId,
-			@NonNull final PickingJobStepId pickingStepId,
-			@Nullable final PickingJobPickFromAlternativeId alternativeId)
-	{
-		final HashMap<HUInfo, I_M_Picking_Job_Step_PickedHU> existingRecordsByPickedHUId = pickedHUs.get(pickingStepId)
-				.stream()
-				.filter(record -> PickingJobPickFromAlternativeId.equals(extractAlternativeId(record), alternativeId))
-				.collect(GuavaCollectors.toHashMapByKey(this::extractPickedHUInfo));
-
-		if (pickedTo != null)
-		{
-			for (final PickingJobStepPickedToHU pickedHU : pickedTo.getActualPickedHUs())
-			{
-				final HUInfo actualPickedHU = pickedHU.getActualPickedHU();
-				I_M_Picking_Job_Step_PickedHU record = existingRecordsByPickedHUId.remove(actualPickedHU);
-				if (record == null)
-				{
-					record = InterfaceWrapperHelper.newInstance(I_M_Picking_Job_Step_PickedHU.class);
-					record.setM_Picking_Job_ID(pickingJobId.getRepoId());
-					record.setM_Picking_Job_Step_ID(pickingStepId.getRepoId());
-					record.setM_Picking_Job_HUAlternative_ID(PickingJobPickFromAlternativeId.toRepoId(alternativeId));
-				}
-
-				record.setPickFrom_HU_ID(pickedHU.getPickFromHUId().getRepoId());
-				record.setPicked_HU_ID(actualPickedHU.getId().getRepoId());
-				record.setPicked_RenderedQRCode(actualPickedHU.getQrCode().toGlobalQRCodeString());
-				record.setC_UOM_ID(pickedHU.getQtyPicked().getUomId().getRepoId());
-				record.setQtyPicked(pickedHU.getQtyPicked().toBigDecimal());
-				record.setCatchWeight(pickedHU.getCatchWeight() != null ? pickedHU.getCatchWeight().toBigDecimal() : null);
-				record.setCatch_UOM_ID(pickedHU.getCatchWeight() != null ? pickedHU.getCatchWeight().getUomId().getRepoId() : -1);
-				InterfaceWrapperHelper.save(record);
-			}
-		}
-
-		InterfaceWrapperHelper.deleteAll(existingRecordsByPickedHUId.values());
-	}
-
-	private void deleteStepsCascade(final Collection<I_M_Picking_Job_Step> steps)
-	{
-		if (steps.isEmpty())
-		{
-			return;
-		}
-
-		final ImmutableSet<PickingJobStepId> pickingJobStepIds = steps.stream().map(step -> PickingJobStepId.ofRepoId(step.getM_Picking_Job_Step_ID())).collect(ImmutableSet.toImmutableSet());
-
-		pickingJobStepIds.forEach(pickingJobStepId -> {
-			InterfaceWrapperHelper.deleteAll(pickingJobStepAlternatives.removeAll(pickingJobStepId));
-			InterfaceWrapperHelper.deleteAll(pickedHUs.removeAll(pickingJobStepId));
-		});
-
-		InterfaceWrapperHelper.deleteAll(steps);
 	}
 
 	public void addAlreadyLoadedFromDB(final I_M_Picking_Job record)
@@ -355,84 +168,30 @@ class PickingJobLoaderAndSaver
 																  + ". Available HU alternatives are: " + pickingJobHUAlternatives));
 	}
 
-	private void loadRecordsFromDB(final Set<PickingJobId> pickingJobIds)
+	private void loadRecords(final Set<PickingJobId> pickingJobIds)
 	{
 		if (pickingJobIds.isEmpty())
 		{
 			return;
 		}
 
-		final List<I_M_Picking_Job> records = InterfaceWrapperHelper.loadByRepoIdAwares(pickingJobIds, I_M_Picking_Job.class);
+		loadRecordsFromDB(pickingJobIds);
 
-		records.forEach(record -> pickingJobs.put(PickingJobId.ofRepoId(record.getM_Picking_Job_ID()), record));
+		final Collection<I_M_Picking_Job> records = pickingJobs.values();
 
-		queryBL.createQueryBuilder(I_M_Picking_Job_HUAlternative.class)
-				.addOnlyActiveRecordsFilter()
-				.addInArrayFilter(I_M_Picking_Job_HUAlternative.COLUMNNAME_M_Picking_Job_ID, pickingJobIds)
-				.create()
-				.stream()
-				.forEach(alt -> pickingJobHUAlternatives.put(PickingJobId.ofRepoId(alt.getM_Picking_Job_ID()), alt));
+		final ImmutableSet<OrderId> salesOrderIds = records.stream().map(PickingJobLoaderAndSaver::extractSalesOrderId).collect(ImmutableSet.toImmutableSet());
+		loadingSupportingServices.warmUpSalesOrderDocumentNosCache(salesOrderIds);
 
-		queryBL.createQueryBuilder(I_M_Picking_Job_Line.class)
-				.addOnlyActiveRecordsFilter()
-				.addInArrayFilter(I_M_Picking_Job_Line.COLUMNNAME_M_Picking_Job_ID, pickingJobIds)
-				.create()
-				.stream()
-				.forEach(line -> pickingJobLines.put(PickingJobId.ofRepoId(line.getM_Picking_Job_ID()), line));
+		final ImmutableSet<BPartnerId> customerIds = records.stream().map(record -> extractDeliveryBPLocationId(record).getBpartnerId()).collect(ImmutableSet.toImmutableSet());
+		loadingSupportingServices.warmUpBPartnerNamesCache(customerIds);
 
-		queryBL.createQueryBuilder(I_M_Picking_Job_Step.class)
-				.addOnlyActiveRecordsFilter()
-				.addInArrayFilter(I_M_Picking_Job_Step.COLUMNNAME_M_Picking_Job_ID, pickingJobIds)
-				.create()
-				.stream()
-				.forEach(step -> pickingJobSteps.put(PickingJobLineId.ofRepoId(step.getM_Picking_Job_Line_ID()), step));
-
-		queryBL.createQueryBuilder(I_M_Picking_Job_Step_HUAlternative.class)
-				.addOnlyActiveRecordsFilter()
-				.addInArrayFilter(I_M_Picking_Job_Step_HUAlternative.COLUMNNAME_M_Picking_Job_ID, pickingJobIds)
-				.create()
-				.stream()
-				.forEach(stepAlt -> pickingJobStepAlternatives.put(extractPickingJobStepId(stepAlt), stepAlt));
-
-		queryBL.createQueryBuilder(I_M_Picking_Job_Step_PickedHU.class)
-				.addOnlyActiveRecordsFilter()
-				.addInArrayFilter(I_M_Picking_Job_Step_PickedHU.COLUMNNAME_M_Picking_Job_ID, pickingJobIds)
-				.create()
-				.stream()
-				.forEach(pickedHU -> pickedHUs.put(PickingJobStepId.ofRepoId(pickedHU.getM_Picking_Job_Step_ID()), pickedHU));
-
-		final PickingJobLoaderSupportingServices loadingSupportingServices = loadingSupportingServicesOrNull();
-		if (loadingSupportingServices != null)
-		{
-			final ImmutableSet<OrderId> salesOrderIds = records.stream().map(PickingJobLoaderAndSaver::extractSalesOrderId).collect(ImmutableSet.toImmutableSet());
-			loadingSupportingServices.warmUpSalesOrderDocumentNosCache(salesOrderIds);
-
-			final ImmutableSet<BPartnerId> customerIds = records.stream().map(record -> extractDeliveryBPLocationId(record).getBpartnerId()).collect(ImmutableSet.toImmutableSet());
-			loadingSupportingServices.warmUpBPartnerNamesCache(customerIds);
-
-			hasLocks.putAll(computePickingJobHasLocks(pickingJobIds));
-		}
-	}
-
-	@NonNull
-	private PickingJobLoaderSupportingServices loadingSupportingServices()
-	{
-		if (_loadingSupportingServices == null)
-		{
-			throw new AdempiereException("No loadingSupportingServices available when saving");
-		}
-		return _loadingSupportingServices;
-	}
-
-	private PickingJobLoaderSupportingServices loadingSupportingServicesOrNull()
-	{
-		return _loadingSupportingServices;
+		hasLocks.putAll(computePickingJobHasLocks(pickingJobIds));
 	}
 
 	private PickingJob loadJob(final I_M_Picking_Job record)
 	{
 		final Optional<PickingSlotIdAndCaption> pickingSlot = Optional.ofNullable(PickingSlotId.ofRepoIdOrNull(record.getM_PickingSlot_ID()))
-				.map(loadingSupportingServices()::getPickingSlotIdAndCaption);
+				.map(loadingSupportingServices::getPickingSlotIdAndCaption);
 
 		final PickingJobId pickingJobId = PickingJobId.ofRepoId(record.getM_Picking_Job_ID());
 
@@ -456,8 +215,6 @@ class PickingJobLoaderAndSaver
 
 	private PickingJobHeader toPickingJobHeader(final I_M_Picking_Job record)
 	{
-		final PickingJobLoaderSupportingServices loadingSupportingServices = loadingSupportingServices();
-
 		final BPartnerLocationId deliveryBPLocationId = extractDeliveryBPLocationId(record);
 		final OrgId orgId = OrgId.ofRepoId(record.getAD_Org_ID());
 
@@ -482,12 +239,12 @@ class PickingJobLoaderAndSaver
 
 		if (luId != null)
 		{
-			final HUQRCode qrCode = loadingSupportingServices().getQRCodeByHUId(luId);
+			final HUQRCode qrCode = loadingSupportingServices.getQRCodeByHUId(luId);
 			return Optional.of(LUPickingTarget.ofExistingHU(luId, qrCode));
 		}
 		else if (luPIId != null)
 		{
-			final String caption = loadingSupportingServices().getPICaption(luPIId);
+			final String caption = loadingSupportingServices.getPICaption(luPIId);
 			return Optional.of(LUPickingTarget.ofPackingInstructions(luPIId, caption));
 		}
 		else
@@ -503,7 +260,7 @@ class PickingJobLoaderAndSaver
 
 		if (tuPIId != null)
 		{
-			final String caption = loadingSupportingServices().getPICaption(tuPIId);
+			final String caption = loadingSupportingServices.getPICaption(tuPIId);
 			return Optional.of(TUPickingTarget.ofPackingInstructions(tuPIId, caption));
 		}
 		else
@@ -524,26 +281,8 @@ class PickingJobLoaderAndSaver
 		return OrderId.ofRepoId(record.getC_Order_ID());
 	}
 
-	private static void updateRecord(final I_M_Picking_Job record, final PickingJob from)
-	{
-		record.setPicking_User_ID(UserId.toRepoId(from.getLockedBy()));
-		record.setM_PickingSlot_ID(from.getPickingSlotId().map(PickingSlotId::getRepoId).orElse(-1));
-
-		final LUPickingTarget pickTarget = from.getLuPickTarget().orElse(null);
-		record.setM_LU_HU_PI_ID(HuPackingInstructionsId.toRepoId(pickTarget != null ? pickTarget.getLuPIId() : null));
-		record.setM_LU_HU_ID(HuId.toRepoId(pickTarget != null ? pickTarget.getLuId() : null));
-
-		final TUPickingTarget tuPickingTarget = from.getTuPickTarget().orElse(null);
-		record.setM_TU_HU_PI_ID(HuPackingInstructionsId.toRepoId(tuPickingTarget != null ? tuPickingTarget.getTuPIId() : null));
-
-		record.setDocStatus(from.getDocStatus().getCode());
-		record.setProcessed(from.getDocStatus().isProcessed());
-	}
-
 	private PickingJobLine loadLine(@NonNull final I_M_Picking_Job_Line record)
 	{
-		final PickingJobLoaderSupportingServices loadingSupportingServices = loadingSupportingServices();
-
 		final ProductId productId = ProductId.ofRepoId(record.getM_Product_ID());
 		final PickingJobLineId pickingJobLineId = PickingJobLineId.ofRepoId(record.getM_Picking_Job_Line_ID());
 
@@ -569,17 +308,6 @@ class PickingJobLoaderAndSaver
 				.build();
 	}
 
-	private static void updateRecord(
-			@NonNull final I_M_Picking_Job_Line record,
-			@NonNull final PickingJobLine line,
-			@NonNull final PickingJobDocStatus docStatus)
-	{
-		final boolean isManuallyClosed = line.isManuallyClosed();
-		record.setIsManuallyClosed(isManuallyClosed);
-
-		record.setProcessed(isManuallyClosed || docStatus.isProcessed());
-	}
-
 	private PickingJobStep loadStep(@NonNull final I_M_Picking_Job_Step record)
 	{
 		final PickingJobStepId pickingJobStepId = PickingJobStepId.ofRepoId(record.getM_Picking_Job_Step_ID());
@@ -602,7 +330,7 @@ class PickingJobLoaderAndSaver
 				//
 				// What?
 				.productId(productId)
-				.productName(loadingSupportingServices().getProductName(productId))
+				.productName(loadingSupportingServices.getProductName(productId))
 				.qtyToPick(Quantitys.create(record.getQtyToPick(), uomId))
 				//
 				// Pick From
@@ -623,7 +351,7 @@ class PickingJobLoaderAndSaver
 				.pickFromKey(PickingJobStepPickFromKey.MAIN)
 				.pickFromLocator(LocatorInfo.builder()
 						.id(locatorId)
-						.caption(loadingSupportingServices().getLocatorName(locatorId))
+						.caption(loadingSupportingServices.getLocatorName(locatorId))
 						.build())
 				.pickFromHU(getHUInfo(pickFromHUId))
 				.pickedTo(loadPickedTo(record))
@@ -649,7 +377,7 @@ class PickingJobLoaderAndSaver
 				.pickFromKey(PickingJobStepPickFromKey.alternative(alternativeId))
 				.pickFromLocator(LocatorInfo.builder()
 						.id(pickFromLocatorId)
-						.caption(loadingSupportingServices().getLocatorName(pickFromLocatorId))
+						.caption(loadingSupportingServices.getLocatorName(pickFromLocatorId))
 						.build())
 				.pickFromHU(getHUInfo(pickFromHUId))
 				.pickedTo(loadPickedTo(record))
@@ -658,13 +386,7 @@ class PickingJobLoaderAndSaver
 
 	private HUQRCode getQRCode(final HuId huId)
 	{
-		return loadingSupportingServices().getQRCodeByHUId(huId);
-	}
-
-	@NonNull
-	private static PickingJobPickFromAlternativeId extractAlternativeId(final I_M_Picking_Job_Step_HUAlternative record)
-	{
-		return PickingJobPickFromAlternativeId.ofRepoId(record.getM_Picking_Job_HUAlternative_ID());
+		return loadingSupportingServices.getQRCodeByHUId(huId);
 	}
 
 	@Nullable
@@ -705,12 +427,6 @@ class PickingJobLoaderAndSaver
 				.filter(record -> PickingJobPickFromAlternativeId.equals(extractAlternativeId(record), alternativeId))
 				.map(this::loadPickedToHU)
 				.collect(ImmutableList.toImmutableList());
-	}
-
-	@Nullable
-	private PickingJobPickFromAlternativeId extractAlternativeId(final I_M_Picking_Job_Step_PickedHU record)
-	{
-		return PickingJobPickFromAlternativeId.ofRepoIdOrNull(record.getM_Picking_Job_HUAlternative_ID());
 	}
 
 	private PickingJobStepPickedToHU loadPickedToHU(@NonNull final I_M_Picking_Job_Step_PickedHU record)
@@ -776,84 +492,6 @@ class PickingJobLoaderAndSaver
 				: Optional.empty();
 	}
 
-	@NonNull
-	private static PickingJobStepId extractPickingJobStepId(final I_M_Picking_Job_Step_HUAlternative record)
-	{
-		return PickingJobStepId.ofRepoId(record.getM_Picking_Job_Step_ID());
-	}
-
-	private static void updateRecord(
-			@NonNull final I_M_Picking_Job_Step record,
-			@NonNull final PickingJobStep from,
-			@NonNull final PickingJobDocStatus docStatus)
-	{
-
-		record.setProcessed(docStatus.isProcessed());
-
-		updateRecord(record, from.getPickFrom(PickingJobStepPickFromKey.MAIN));
-	}
-
-	private static void updateRecord(
-			@NonNull final I_M_Picking_Job_Step record,
-			@NonNull final PickingJobStepPickFrom mainPickFrom)
-	{
-		record.setPickFrom_Warehouse_ID(mainPickFrom.getPickFromWarehouseId().getRepoId());
-		record.setPickFrom_Locator_ID(mainPickFrom.getPickFromLocatorId().getRepoId());
-		record.setPickFrom_HU_ID(mainPickFrom.getPickFromHUId().getRepoId());
-		updateRecord(record, mainPickFrom.getPickedTo());
-	}
-
-	private static void updateRecord(
-			@NonNull final I_M_Picking_Job_Step record,
-			@Nullable final PickingJobStepPickedTo pickedTo)
-	{
-		final BigDecimal qtyRejectedBD;
-		final String rejectReason;
-		if (pickedTo != null)
-		{
-			qtyRejectedBD = pickedTo.getQtyRejected() != null ? pickedTo.getQtyRejected().toBigDecimal() : BigDecimal.ZERO;
-			rejectReason = pickedTo.getQtyRejected() != null ? pickedTo.getQtyRejected().getReasonCode().getCode() : null;
-		}
-		else
-		{
-			qtyRejectedBD = BigDecimal.ZERO;
-			rejectReason = null;
-		}
-
-		record.setQtyRejectedToPick(qtyRejectedBD);
-		record.setRejectReason(rejectReason);
-	}
-
-	private static void updateRecord(final I_M_Picking_Job_Step_HUAlternative existingRecord, final PickingJobStepPickFrom pickFrom)
-	{
-		existingRecord.setPickFrom_HU_ID(pickFrom.getPickFromHUId().getRepoId());
-		updateRecord(existingRecord, pickFrom.getPickedTo());
-	}
-
-	private static void updateRecord(final I_M_Picking_Job_Step_HUAlternative existingRecord, final PickingJobStepPickedTo pickedTo)
-	{
-		final UomId uomId;
-		final BigDecimal qtyRejectedBD;
-		final String rejectReason;
-		if (pickedTo != null && pickedTo.getQtyRejected() != null)
-		{
-			final QtyRejectedWithReason qtyRejected = pickedTo.getQtyRejected();
-			uomId = qtyRejected.toQuantity().getUomId();
-			qtyRejectedBD = qtyRejected.toBigDecimal();
-			rejectReason = qtyRejected.getReasonCode().getCode();
-		}
-		else
-		{
-			uomId = null;
-			qtyRejectedBD = BigDecimal.ZERO;
-			rejectReason = null;
-		}
-
-		existingRecord.setC_UOM_ID(UomId.toRepoId(uomId));
-		existingRecord.setQtyRejectedToPick(qtyRejectedBD);
-		existingRecord.setRejectReason(rejectReason);
-	}
-
 	private PickingJobPickFromAlternative loadPickFromAlternative(final I_M_Picking_Job_HUAlternative record)
 	{
 		final LocatorId locatorId = LocatorId.ofRepoId(record.getPickFrom_Warehouse_ID(), record.getPickFrom_Locator_ID());
@@ -863,7 +501,7 @@ class PickingJobLoaderAndSaver
 				.id(PickingJobPickFromAlternativeId.ofRepoId(record.getM_Picking_Job_HUAlternative_ID()))
 				.locatorInfo(LocatorInfo.builder()
 						.id(locatorId)
-						.caption(loadingSupportingServices().getLocatorName(locatorId))
+						.caption(loadingSupportingServices.getLocatorName(locatorId))
 						.build())
 				.pickFromHU(getHUInfo(pickFromHUId))
 				.productId(ProductId.ofRepoId(record.getM_Product_ID()))
@@ -952,7 +590,7 @@ class PickingJobLoaderAndSaver
 
 		final ImmutableSetMultimap<PickingJobId, ShipmentScheduleId> shipmentScheduleIdsByPickingJobId = getShipmentScheduleIds(pickingJobIds);
 
-		final SetMultimap<ShipmentScheduleId, ExistingLockInfo> existingLocks = loadingSupportingServices().getLocks(shipmentScheduleIdsByPickingJobId.values());
+		final SetMultimap<ShipmentScheduleId, ExistingLockInfo> existingLocks = loadingSupportingServices.getLocks(shipmentScheduleIdsByPickingJobId.values());
 
 		final ImmutableMap.Builder<PickingJobId, Boolean> result = ImmutableMap.builder();
 		for (final PickingJobId pickingJobId : pickingJobIds)
@@ -981,7 +619,7 @@ class PickingJobLoaderAndSaver
 	private PickingUnit computePickingUnit(@Nullable final UomId catchUomId, @NonNull final HUPIItemProduct packingInfo)
 	{
 		// If catch weight, always pick at CU level because user has to weight the products
-		if (!loadingSupportingServices().isCatchWeightTUPickingEnabled() && catchUomId != null)
+		if (!loadingSupportingServices.isCatchWeightTUPickingEnabled() && catchUomId != null)
 		{
 			return PickingUnit.CU;
 		}
