+++ backend/de.metas.picking.rest-api/src/main/java/de/metas/picking/workflow/PickingJobRestService.java
@@ -194,4 +194,9 @@ public class PickingJobRestService
 	{
 		return mobileUIPickingUserProfileRepository.getProfile().isAllowNewTU();
 	}
+
+	public boolean isShowPromptWhenOverPicking()
+	{
+		return mobileUIPickingUserProfileRepository.getProfile().isShowConfirmationPromptWhenOverPick();
+	}
 }
