+++ misc/services/mobile-webui/mobile-webui-frontend/package.json
@@ -29,6 +29,7 @@
     "react-router": "^5.2.1",
     "react-router-dom": "^5.3.0",
     "react-scripts": "4.0.3",
+    "react-toastify": "7.0.0",
     "redux-localstorage-simple": "^2.4.1",
     "reselect": "^4.0.0",
     "sockjs-client": "^1.5.2",
