+++ misc/services/mobile-webui/mobile-webui-frontend/src/components/ScreenToaster.jsx
@@ -1,38 +1,20 @@
-import React, { useEffect } from 'react';
-import toast, { Toaster, useToasterStore } from 'react-hot-toast';
-
-const TOASTS_LIMIT = 1;
+import React from 'react';
+import { ToastContainer } from 'react-toastify';
 
 const ScreenToaster = () => {
-  //
-  // Limit the number of toasts displayed on screen
-  // (thx to https://github.com/timolins/react-hot-toast/issues/31#issuecomment-803359550)
-  if (TOASTS_LIMIT > 0) {
-    const { toasts } = useToasterStore();
-    useEffect(() => {
-      toasts
-        .filter((t) => t.visible)
-        .filter((_, i) => i >= TOASTS_LIMIT)
-        .forEach((t) => toast.dismiss(t.id));
-    }, [toasts]);
-  }
-
   return (
-    <Toaster
+    <ToastContainer
       position="bottom-center"
-      containerClassName="app-toaster"
-      toastOptions={{
-        success: {
-          style: {
-            background: 'green',
-          },
-        },
-        error: {
-          style: {
-            background: 'red',
-          },
-        },
-      }}
+      autoClose={5000}
+      hideProgressBar={true}
+      newestOnTop={true}
+      closeOnClick
+      rtl={false}
+      pauseOnFocusLoss
+      draggable
+      pauseOnHover
+      theme="dark"
+      style={{ marginBottom: '3rem' }}
     />
   );
 };
