+++ misc/services/mobile-webui/mobile-webui-frontend/src/containers/activities/picking/PickStepScanScreen.jsx
@@ -1,10 +1,15 @@
-import React, { useEffect } from 'react';
+import React, { useCallback, useEffect } from 'react';
 import { useHistory, useRouteMatch } from 'react-router-dom';
 import { shallowEqual, useDispatch, useSelector } from 'react-redux';
 
 import { trl } from '../../../utils/translations';
 import { pushHeaderEntry } from '../../../actions/HeaderActions';
-import { getActivityById, getQtyRejectedReasonsFromActivity, getStepById } from '../../../reducers/wfProcesses';
+import {
+  getActivityById,
+  getLineById,
+  getQtyRejectedReasonsFromActivity,
+  getStepById,
+} from '../../../reducers/wfProcesses';
 import { toastError } from '../../../utils/toast';
 import { getPickFromForStep, getQtyToPickForStep } from '../../../utils/picking';
 import { postStepPicked } from '../../../api/picking';
@@ -12,6 +17,7 @@ import { postStepPicked } from '../../../api/picking';
 import ScanHUAndGetQtyComponent from '../../../components/ScanHUAndGetQtyComponent';
 import { toQRCodeString } from '../../../utils/qrCode/hu';
 import { updateWFProcess } from '../../../actions/WorkflowActions';
+import { toNumberOrZero } from '../../../utils/numberUtils';
 
 const PickStepScanScreen = () => {
   const {
@@ -19,9 +25,20 @@ const PickStepScanScreen = () => {
     params: { workflowId: wfProcessId, activityId, lineId, stepId, altStepId },
   } = useRouteMatch();
 
-  const { eligibleQRCode, qtyToPick, uom, qtyRejectedReasons } = useSelector(
-    (state) => getPropsFromState({ state, wfProcessId, activityId, lineId, stepId, altStepId }),
-    shallowEqual
+  const { eligibleQRCode, qtyToPick, uom, qtyRejectedReasons, qtyRemainingToPick, isShowPromptWhenOverPicking } =
+    useSelector(
+      (state) => getPropsFromState({ state, wfProcessId, activityId, lineId, stepId, altStepId }),
+      shallowEqual
+    );
+
+  const getConfirmationPromptForQty = useCallback(
+    (qtyInput) => {
+      if (qtyRemainingToPick !== undefined && toNumberOrZero(qtyInput) > qtyRemainingToPick) {
+        return trl('activities.picking.overPickConfirmationPrompt');
+      }
+      return undefined;
+    },
+    [qtyRemainingToPick]
   );
 
   const dispatch = useDispatch();
@@ -65,6 +82,7 @@ const PickStepScanScreen = () => {
       uom={uom}
       qtyRejectedReasons={qtyRejectedReasons}
       //
+      getConfirmationPromptForQty={isShowPromptWhenOverPicking ? getConfirmationPromptForQty : undefined}
       onResult={onResult}
     />
   );
@@ -74,6 +92,7 @@ const getPropsFromState = ({ state, wfProcessId, activityId, lineId, stepId, alt
   const activity = getActivityById(state, wfProcessId, activityId);
   const qtyRejectedReasons = getQtyRejectedReasonsFromActivity(activity);
 
+  const lineProps = getLineById(state, wfProcessId, activityId, lineId);
   const stepProps = getStepById(state, wfProcessId, activityId, lineId, stepId);
   const eligibleQRCode = toQRCodeString(getPickFromForStep({ stepProps, altStepId }).huQRCode);
   const qtyToPick = getQtyToPickForStep({ stepProps, altStepId });
@@ -83,6 +102,8 @@ const getPropsFromState = ({ state, wfProcessId, activityId, lineId, stepId, alt
     qtyToPick,
     uom: stepProps.uom,
     qtyRejectedReasons,
+    qtyRemainingToPick: lineProps.qtyRemainingToPick,
+    isShowPromptWhenOverPicking: activity?.dataStored?.isShowPromptWhenOverPicking,
   };
 };
