+++ misc/services/mobile-webui/mobile-webui-frontend/src/utils/translations_de.js
@@ -115,6 +115,7 @@ const translations = {
         Current: 'Current',
         CloseTarget: 'Close',
       },
+      overPickConfirmationPrompt: 'Möchten Sie wirklich mehr kommissionieren als beauftragt wurde?',
     },
     distribution: {
       DistributionLine: 'Bereitstellung Zeile',
