+++ misc/services/mobile-webui/mobile-webui-frontend/src/utils/translations_en.js
@@ -115,6 +115,7 @@ const translations = {
         Current: 'Current',
         CloseTarget: 'Close',
       },
+      overPickConfirmationPrompt: 'Do you really want to pick more than was ordered?',
     },
     distribution: {
       DistributionLine: 'Distribution Line',
